New data, discrepancies
Added measurements from today, there is a discrepancy in the beat note
readout and the sensitivity expected from sensitivity limited to PD
noise. Hopefully this is just a calibration error. The TFs don’t line
up either. I added a section to plot the readout at the 02/20 control
signals, and these show the same discrepancy. Also there seems to be a
discrepancy in the predicted error signal slopes even though the
transmission and coupling now are converging to reasonable values.