Added intensity noise, changed 02 loop
The discrepancy in the beat note and the PD noise seems to be real in
the sense that there is another noise source unaccounted for in our
noise budget. It seems to be coming from the light, as the noise floor
rises with light hitting the PD.  I added a contribution due to the
intensity noise of the laser itself, but it looks to be below the shot
noise limit.  The TFs still don’t match up. If we assume the
discrepancy is due to the servo, which all measurements indicate so
far, this should not affect the noise budget within the measurement
band.